Moved old sketches to separate directory, created file for CIDs, code rework
Continuing to revise the newest version of the sketch, created a directory to put all the old ones in for preservation's sake, and starting a new file(s) to make a sort of table showing the CID of everything. The tables showing CIDs will likely be structured in a way that others can contribute to.